requirements.txt
# Core ML/AI Dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
sentence-transformers>=2.2.0
clip-by-openai>=1.0
openai>=1.0.0

# Qwen MLLM Integration
qwen-vl-utils>=0.0.1
modelscope>=1.9.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=9.5.0
albumentations>=1.3.0

# SAM and Segmentation
segment-anything>=1.0
ultralytics>=8.0.0
detectron2>=0.6.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0

# Web Framework
flask>=2.3.0
gradio>=4.0.0
werkzeug>=2.3.0

# Similarity digitizer dependencies
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
